Restructuring v-pool test directory
